Refactor calculator to use int instead of double
Updated project files and classes to transition from double to int for calculations and memory management. Adjusted methods in Calculator, Memory, and MemoryItem classes. Modified tests to reflect these changes and updated the WinForms UI for memory item handling. Added project references for WinFormsApp1.

+++ CalculatorLibrary/IManagable.cs
@@ -8,7 +8,7 @@ namespace CalculatorLibrary
 {
     public interface IManagable
     {
-        void Add(double num);
-        void Sub(double num);
+        void Add(int num);
+        void Sub(int num);
     }
 }
